requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
